+++ pkg/innerring/innerring.go
@@ -936,6 +936,7 @@ func New(ctx context.Context, log *zap.Logger, cfg *viper.Viper, errChan chan<- 
 		ContainerClient: cnrClient,
 		NeoFSIDClient:   neofsIDClient,
 		NetworkState:    server.netmapClient,
+		MetaEnabled:     cfg.GetBool("experimental.chain_meta_data"),
 	})
 	if err != nil {
 		return nil, err